+++ packages/evokit-image/CHANGELOG.md
@@ -1,3 +1,5 @@
+[create_theme]: docs/base/theme
+
 # EvoKit - Image
 
 [![](https://img.shields.io/npm/v/evokit-image.svg)](https://www.npmjs.com/package/evokit-image)
@@ -5,6 +7,17 @@
 ---
 
 ### CHANGELOG
+
+#### 3.0.2 *(11.07.2019)*
+
+- **Added** new advanced border props:
+    - `image-border-top`
+    - `image-border-right`
+    - `image-border-bottom`
+    - `image-border-left`
+    - `image-border-tb`
+    - `image-border-lr`
+- Please [update theme][create_theme]
 
 #### 3.0.1 *(25.06.2019)*